liqoctl: rename package inbound to inband
